# app/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Notification Service",
    "description": "Mass SMS/Email dispatch with priority routing, durable queues, and business-level exactly-once delivery.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/notifications": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Create Notifications",
        "operationId": "create_notifications_api_v1_notifications_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateNotificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notification_id}": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Get Notification",
        "operationId": "get_notification_api_v1_notifications__notification_id__get",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationWithHistory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subscribers/{subscriber_id}/notifications": {
      "get": {
        "tags": [
          "subscribers"
        ],
        "summary": "List Subscriber Notifications",
        "operationId": "list_subscriber_notifications_api_v1_subscribers__subscriber_id__notifications_get",
        "parameters": [
          {
            "name": "subscriber_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subscriber Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "queued",
                    "sent",
                    "delivered",
                    "rejected"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationWithHistory"
                  },
                  "title": "Response List Subscriber Notifications Api V1 Subscribers  Subscriber Id  Notifications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateNotificationRequest": {
        "properties": {
          "channel": {
            "type": "string",
            "enum": [
              "sms",
              "email"
            ],
            "title": "Channel"
          },
          "type": {
            "type": "string",
            "enum": [
              "transactional",
              "marketing"
            ],
            "title": "Type"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "recipient_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recipient Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "channel",
          "type",
          "message",
          "recipient_ids"
        ],
        "title": "CreateNotificationRequest"
      },
      "CreateNotificationResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Batch Id"
          },
          "channel": {
            "type": "string",
            "enum": [
              "sms",
              "email"
            ],
            "title": "Channel"
          },
          "type": {
            "type": "string",
            "enum": [
              "transactional",
              "marketing"
            ],
            "title": "Type"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "recipients": {
            "items": {
              "$ref": "#/components/schemas/RecipientStatus"
            },
            "type": "array",
            "title": "Recipients"
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "channel",
          "type",
          "total",
          "recipients"
        ],
        "title": "CreateNotificationResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthComponent": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "HealthComponent"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "degraded"
            ],
            "title": "Status"
          },
          "components": {
            "additionalProperties": {
              "$ref": "#/components/schemas/HealthComponent"
            },
            "type": "object",
            "title": "Components"
          }
        },
        "type": "object",
        "required": [
          "status",
          "components"
        ],
        "title": "HealthResponse"
      },
      "NotificationWithHistory": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Batch Id"
          },
          "subscriber_id": {
            "type": "string",
            "title": "Subscriber Id"
          },
          "channel": {
            "type": "string",
            "enum": [
              "sms",
              "email"
            ],
            "title": "Channel"
          },
          "type": {
            "type": "string",
            "enum": [
              "transactional",
              "marketing"
            ],
            "title": "Type"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "sent",
              "delivered",
              "rejected"
            ],
            "title": "Status"
          },
          "provider_message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Message Id"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "sent_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          },
          "delivered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered At"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/StatusEventOut"
            },
            "type": "array",
            "title": "History",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "batch_id",
          "subscriber_id",
          "channel",
          "type",
          "status",
          "provider_message_id",
          "retry_count",
          "last_error",
          "created_at",
          "sent_at",
          "delivered_at"
        ],
        "title": "NotificationWithHistory"
      },
      "RecipientStatus": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "subscriber_id": {
            "type": "string",
            "title": "Subscriber Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "sent",
              "delivered",
              "rejected"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subscriber_id",
          "status"
        ],
        "title": "RecipientStatus"
      },
      "StatusEventOut": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "sent",
              "delivered",
              "rejected"
            ],
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "status",
          "detail",
          "created_at"
        ],
        "title": "StatusEventOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}